+++ baitstools.rb
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 #-----------------------------------------------------------------------------------------------
 # baitstools
-BAITSTOOLSVER = "1.5.0"
+BAITSTOOLSVER = "1.6.0"
 # Michael G. Campana, 2017-2019
 # Smithsonian Conservation Biology Institute
 #-----------------------------------------------------------------------------------------------
@@ -111,6 +111,11 @@ class Parser
 		args.default_files = [] # Default files to be written
 		args.rng = srand # Random number seed
 		args.gzip = false # Gzip output files
+		args.fiveprime = "" # Sequence to addend to 5' end
+		args.threeprime = "" # Sequence to addend to 3' end
+		args.fillin = "" # Motif to fill-in short baits
+		args.fillin_switch = false # Switch to turn off if filling in
+		args.noaddenda = false # Exclude 5' and 3' addended sequences from QC calculations
 		opt_parser = OptionParser.new do |opts|
 			if algorithms.include?(args.algorithm) # Process correct commands
 				opts.banner = "Command-line usage: ruby baitstools.rb "+args.algorithm+" [options]"
@@ -297,8 +302,8 @@ class Parser
 				opts.on("-N", "--noNs", "Exclude bait sequences with Ns") do
 					args.no_Ns = true
 				end
-				opts.on("-C", "--collapse", "Collapse ambiguities to a single nucleotide") do
-					args.collapse_ambiguities = true
+				opts.on("--noaddenda", "Exclude 5' and 3' addended sequences from bait parameter calculations") do
+					args.noaddenda = true
 				end
 				opts.on("-n","--mingc [VALUE]", Float, "Minimum GC content in % (Default = 30.0)") do |mgc|
 					args.mingc = mgc if mgc != nil
@@ -380,6 +385,9 @@ class Parser
 						args.ncbi = true
 					end
 				end
+				opts.on("-C", "--collapse", "Collapse ambiguities to a single nucleotide") do
+					args.collapse_ambiguities = true
+				end
 				opts.on("-Y","--rna", "Output baits as RNA rather than DNA") do
 					args.rna = true
 				end
@@ -391,6 +399,18 @@ class Parser
 				end
 				opts.on("-G", "--gaps [VALUE]", String, "Strategy to handle sequence gaps (-) (include, exclude, or extend) (Default = include)") do |gap|
 					args.gaps = gap if gap != nil
+				end
+				opts.on("-5", "--5prime [VALUE]", String, "Sequence to addend to 5' end of baits") do |fivepr|
+					args.fiveprime = fivepr if fivepr != nil
+				end
+				opts.on("-3", "--3prime [VALUE]", String, "Sequence to addend to 3' end of baits") do |threepr|
+					args.threeprime = threepr if threepr != nil
+				end
+				opts.on("--fillin [VALUE]", String, "Fill in baits shorter than requested length with specified sequence repeat motif") do |fillin|
+					if fillin != nil
+						args.fillin = fillin
+						args.fillin_switch = true
+					end
 				end
 				opts.on("-X", "--threads [VALUE]", Integer, "Number of threads (Default = 1)") do |thr|
 					args.threads = thr if thr != nil
@@ -897,7 +917,33 @@ begin
 		t = gets.chomp.upcase
 		if t == "Y" or t == "YES"
 			$options.rc = true
-		end 
+		end
+		print "Addend a sequence to 5' end of baits? (y/n)\n"
+		t = gets.chomp.upcase
+		if t == "Y" or t == "YES"
+			print "Enter 5' sequence.\n"
+			$options.fiveprime = gets.chomp
+		end
+		print "Addend a sequence to 3' end of baits? (y/n)\n"
+		t = gets.chomp.upcase
+		if t == "Y" or t == "YES"
+			print "Enter 3' sequence.\n"
+			$options.threeprime = gets.chomp
+		end
+		unless $options.fiveprime == "" && $options.threeprime == ""
+			print "Exclude sequence addenda from bait filtration parameter calculations? (y/n)\n"
+			t = gets.chomp.upcase
+			if t == "Y" or t == "YES"
+				$options.noaddenda = true
+			end
+		end
+		print "Fill in incomplete baits with a specified sequence repeat motif? (y/n)\n"
+		t = gets.chomp.upcase
+		if t == "Y" or t == "YES"
+			print "Enter sequence motif.\n"
+			$options.fillin = gets.chomp
+			$options.fillin_switch = true unless $options.fillin == ""
+		end
 		print "Filter by minimum GC content? (y/n)\n"
 		t = gets.chomp.upcase
 		if t == "Y" or t == "YES"
@@ -1062,6 +1108,8 @@ begin
 	print "** Filtration options:" + cmdline[1] + " **\n" # filtered line always starts with a space if present
 	setup_output
 	build_fq_hash # Build FASTQ quality translation hash
+	build_rc_hash # Build reverse complementation hash
+	build_ambig_hash # Build ambiguity hash
 	write_file(".log.txt", "Parsed commands: " + cmdline[0] + cmdline[1] + "\n") if $options.log
 	case $options.algorithm
 	when "aln2baits"