+++ CHANGELOG.md
@@ -20,6 +20,13 @@ Contact: campanam@si.edu
 [Deprecated](#deprecated)  
 
 ## aln2baits  
+### Version 1.6.0  
+5' and 3' addenda  
+Sequence motif fill-in  
+Correction of bugs/improvement of reversecomp/RNA qualities and ordering  
+Used thread bug fix  
+Collapsed baits output  
+
 ### Version 1.4.0  
 Circular sequence handling  
 
@@ -112,6 +119,16 @@ Version constant added to header
 Preliminary script to generate baits from an annotation file and a reference sequence  
 
 ## baitslib  
+### Version 1.6.0  
+5' and 3' addenda  
+Sequence motif fill-in  
+Correction of bugs/improvement of reversecomp/RNA qualities and ordering  
+Numerous new methods for improved handling  
+Used thread bug fix  
+read_fasta bug fix  
+Revision of ambiguity handling code  
+Collapsed baits output  
+
 ### Version 1.5.0  
 Read/write gzip files  
 
@@ -216,6 +233,12 @@ New method write_probes handles basic output
 filter_probes definition removed into separate script for access by other scripts  
 
 ## baitstools  
+### Version 1.6.0  
+5' and 3' addenda  
+Sequence motif fill-in  
+--collapse moved to general options  
+Ambiguities and reverse complement hashes generated at start-up  
+
 ### Version 1.5.0  
 Read/write gzip files  
 
@@ -350,6 +373,9 @@ The word 'probe' changed to 'baits' in all instances for clarity
 Set default for tiling offset as 20 bp (from 60 for select_snps and 25 for tile_probes)  
 
 ## baitstoolsgui  
+### Version 1.6.0  
+Sequence addenda/fill-in handling added  
+
 ### Version 1.5.0  
 Read/write gzip files  
 
@@ -443,6 +469,9 @@ Popvar @alleles replaces @pnuc and @qnuc, which were otherwise unused
 Preliminary script to turn a coordinates table and a reference sequence into baits  
 
 ## blast2baits  
+### Version 1.6.0  
+Correction of bugs/improvement of reversecomp/RNA qualities and ordering  
+
 ### Version 1.5.0  
 Read/write gzip files  
 
@@ -466,6 +495,12 @@ Region tiling code transferred to baitslib
 Preliminary script to generate baits from BLAST hit tables and a reference sequence  
 
 ## checkbaits  
+5' and 3' addenda  
+Sequence motif fill-in  
+Code-update and bug fixes  
+Used thread bug fix  
+Collapsed baits output  
+
 ### Version 1.1.0  
 Reworking to use new temp file methods  
 
@@ -503,6 +538,9 @@ Versioning added to header of script
 RVM default bug fixed  
 
 ## pyrad2baits  
+### Version 1.6.0  
+Updated to use $ambig_hash  
+
 ### Version 1.5.0  
 Read/write gzip files  
 
@@ -544,6 +582,13 @@ Fixed bug in that reread reference sequence every time snps_to_baits run in stac
 Preliminary script to turn a Stacks summary tsv file and a reference sequence into baits  
 
 ## tilebaits  
+### Version 1.6.0  
+5' and 3' addenda  
+Sequence motif fill-in  
+Correction of bugs/improvement of reversecomp/RNA qualities and ordering  
+Used thread bug fix  
+Collapsed baits output  
+
 ### Version 1.4.0  
 Baits can extend multiple times around ultra-short reference sequences  
 Fixed bug in quality scores for FASTQ files